# bin/spec/openapi.yaml
openapi: '3.0.2'
servers:
  - url: 'https://hedbot.uc.r.appspot.com/eliza/api/'
    description: Development Server
info:
  title: Eliza API
  version: '1.0'
  description: |
    Supports control and retrieval over Eliza programs by Elli

    # Introduction
    This API is documented in **OpenAPI format**
    It is **extended** to support features of [ReDoc](https://github.com/Redocly/redoc) documentation.

    # Authentication

    The Eliza REST API supports HTTP Bearer (JWT)  Authentication as described in [spec] (https://deadlink.hms.com/)

    For this sample, use the JWT Token displayed when you click the "Authorize" button. Cut/past into the
    value field and the token will be used for subsequent calls. **All** of the API calls require authentication/authorization.
...
components:
  schemas:
...
    ProgramDef:
      type: object
      properties:
        _links:
          type: object
          properties:
            self:
              $ref: '#/components/schemas/Link'
        _associations:
          type: object
          properties:
            client:
              $ref: '#/components/schemas/Association'
            solution:
              $ref: '#/components/schemas/Association'
        id:
          type: string
        displayName:
          type: string
        description:
          type: string
        isTransactional:
          type: boolean
        _embedded:
          type: object
          properties:
            programOutreaches:
              type: array
              items:
                $ref: '#/components/schemas/ResourceRef'
...
    MemberInfoSubmission:
      type: array
      items:
        type: object
        properties:
          transactionXid:
            type: string
          memberXid:
            type: string
          groupXid:
            type: string
          batchXid:
            type: string
          firstName:
            type: string
            x-pii: true
          lastName:
            type: string
            x-pii: true
          dob:
            type: string
            x-pii: true
          gender:
            type: string
            x-pii: true
          preferredLanguage:
            type: string
            x-pii: true
          email:
            type: string
            x-pii: true
          primaryPhone:
            type: string
            x-pii: true
          secondaryPhone:
            type: string
            x-pii: true
          address1:
            type: string
            x-pii: true
          address2:
            type: string
            x-pii: true
          city:
            type: string
            x-pii: true
          state:
            type: string
            x-pii: true
          postalCode:
            type: string
            x-pii: true
          lineOfBusiness:
            type: string
          allowedChannels:
            type: object
            properties:
              telephone:
                type: boolean
              messaging:
                type: boolean
              email:
                type: string
          allowedContactDays:
            type: object
            properties:
              monday:
                type: boolean
              tuesday:
                type: boolean
              wednesday:
                type: boolean
              thursday:
                type: boolean
              friday:
                type: boolean
              saturday:
                type: boolean
              sunday:
                type: boolean
          test:
            type: boolean
        additionalProperties:
          type: string
    MemberInfoSubmissionResponse:
      type: array
      items:
        type: object
        properties:
          transactionXid:
            type: string
          memberXid:
            type: string
          dataRecordId:
            type: string
...
    Link:
      type: object
      properties:
        href:
          type: string
        rel:
          type: string
    Association:
      type: object
      properties:
        href:
          type: string
        rel:
          type: string
        id:
          type: string
        displayName:
          type: string
    ResourceRef:
      type: object
      properties:
        _links:
          type: object
          properties:
            self:
              $ref: '#/components/schemas/Link'
        id:
          type: string
        displayName:
          type: string
    ErrorResponse:
      type: object
      properties:
        _links:
          type: object
          properties:
            self:
              $ref: '#/components/schemas/Link'
        errorMessage:
          type: string
        errorDetail:
          type: object
  parameters:
...
    clientId:
      name: clientId
      in: path
      required: true
      description:  Client Id (Eliza Internal)
      schema:
        type: string
    programId:
      name: programId
      in: path
      required: true
      description:  Program Id
      schema:
        type: string
...
  securitySchemes:
    name:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /clients/{clientId}/programs/{programId}:
    get:
      summary: Program Detail
      description: Program Detail
      tags:
        - Navigation
      parameters:
        - $ref: '#/components/parameters/clientId'
        - $ref: '#/components/parameters/programId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgramDef'

    post:
      summary: Member Submission
      description:  |
        # DataRecord/Entity Submission

        **Proposal**

        Support the submission of 1 or more 'Data Records' to the system.

        Data Records submitted as a collection/array of JSON strucures that capture all of the
        required and optional elements needed to enroll a member into a program

        The API implementation would "pass through" the request to an internal "micro-service"
        that would receive the records, assign internal Ids, and return those Ids to the caller. All that
        is strictly required is that an ID is assigned, the request is queued, and there is a reasonable (high)
        assurance that once queued, a data record with that Id will be available to query in the future.

        NOTE: that the structure is "open ended" in that it can support any number
        additional named fields ('F' fields).

        NOTE: This specific API method replaces part of (but not all of) the Web Services API - the part that
        ingests member records/entities.

        NOTE:  The batchId supplied by the client is renamed to batchXid to clearly distinguish as eXternal and
        distinct from the internal batchId

        NOTE: The member number supplied by the client is renamed memberXid to clearly distinguish it as eXternal
        and distinct from any internal/generic memberId

        NOTE: The response code is 202-Accepted - meaning that unless there's an error on receipt, the
        record is accepted and returns a dataRecordId. That does not mean the record will be processed. It may
        fail and subsequent status queries against that dataRecordId should describe the failure.

        # Open Questions/Issues:

        1) Does the implemmentation capture all of the Web Services ingestion use cases?
        2) Assumes the entity will either trigger creation of a new batch (if first record) or be added to programs
        batch as appropriate - same as current web service. Correct?
        3) Added secondary phone. Not sure it that's useful/desirable/required, or otherwise adventageous


      tags:
        - PHMi MVP Ingestion
      parameters:
        - $ref: '#/components/parameters/clientId'
      requestBody:
        description: Member Data Records (Collection)
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberInfoSubmission'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberInfoSubmissionResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
...